Throw error instead of emit outside of stream
Some errors are going unreported because they are not being thrown. Gulp docs indicate they should be thrown if not in stream.

In my case the compile error "You are trying to use "hot-reload". vue-hot-reload-api is missing. ..." was causing gulp to die silently. The change above properly reports the error.